제성수정 리뷰 목록불러오기 수정필요

+++ project04/src/main/java/com/ex/repository/ReviewRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
-    List<ReviewEntity> findByProduct_Id(Long productId); // 상품 ID로 리뷰 검색
+    List<ReviewEntity> findByProduct_Id(Integer productId); // 상품 ID로 리뷰 검색
 }
